Add failing test for `isinstance(<union>)`
Summary:
Having these two equivalent tests side-by-side is going to be useful
for fixing the distribute-narrow-op-over-tuple logic, which is not going
to work well in its current form for attributes.

I figured it probably already worked when I wrote this test, but it turns
out I found a missing feature, we don't currently handle unions in
`isinstance` which we should (at runtime it works, tuples and unions of
simple classes behave equivalently).

I'll probably wind up implementing support for this in the process of
eliminating the use of distribute

Reviewed By: ndmitchell

Differential Revision: D72645994

fbshipit-source-id: 7459ba9577d6d1205b0b3a1f2db449d08f91c7bf