Introduce `as_decomposed_tuple` helper
Summary:
The biggest blocker remaining to actually producing attribute narrows in the narrowing
code is that at the moment Pyrefly is not capable of applying narrowing operations to
a tuple directly.

Instead, the implementation relies on constructing a `NarrowOps` on the fly; this won't
work in a world with attribute narrowing because the composeable `NarrowOps` tree
needs to operate at the `TypeInfo` level, whereas the disribute operation is getting
called inside of the `AtomicNarrowOp` handler that needs to operate at the type level.

I need to undo this; a first step is to make the case analysis match directly on
the condition that we actually care about which is the right type being a tuple,
rather than matching on the result of the distribute operation we need to eliminate.

Reviewed By: ndmitchell

Differential Revision: D72645995

fbshipit-source-id: 7e36aa2574337922646765ab98c91b27e1fadf76