Kill distribute-over-tuple
Summary:
This commit eliminates the distribute-over-tuple logic from attribute
narrowing, which is big progress because that operation was relying
on intermingling the control-flow-relevant `And` and `Or` "tree" part
of narrowing (which has to be `TypeInfo` level so that attribute narrows
can propagate through control flow) with the atomic part of narrowing
(which can and should be kept `Type` level for modularity).

As an added benefit, we can now eliminate `NarrowVal::Type`, which was
only needed due to this solve-time synthesis of new `NarrowOps` values
and is no longer used.

Reviewed By: ndmitchell

Differential Revision: D72645999

fbshipit-source-id: 535fc0913ba698574afdb3bd5e0c847177e0c09f