Remove no-longer-needed `NarrowVal` abstraction
Summary:
It's nice to get rid of one layer of abstraction in the process of
attribute narrowing support, since we had to add one.

We no longer need this abstraction because we are handling the tuple case
of `isinstance` correctly in a direct way rather than by synthesizing
a `NarrowOps` at solve time; as a result, `NarrowOps` is now a binding-phase-only
data structure and always has expressions rather than types for the narrow values.

Reviewed By: ndmitchell

Differential Revision: D72645996

fbshipit-source-id: 77a2e420fae7a6032cc7e60c6a48b9d553e62919