Run only one transaction per validation
Summary:
Instead of running a transaction for each config, we only need to run it once.

This is obvious for on-disk changes, since the transaction we run is exactly the same (note that we don't really need to run with open file handles there).

For in-memory validation, we just need to do a little bit of extra work to split in into 3 phases. First we compute all handles to run transactions on and invalidate, then we run the transaction, and we finally obtain the errors to print for each config.

To avoid duplicating some code with handles and diagnostics, I also refactored them into helper functions.

Reviewed By: ndmitchell

Differential Revision: D72662021

fbshipit-source-id: f6a59f99e949cc0fc45d9b4ec6d3de441eada107