Check in pyrefly_wasm.js
Summary:
I had to remove `pyrefly_wasm.js` from `.gitignore` in D72569996 to make sure we can move the empty interface with the directory wide rename. I originally thought about adding it back to the `.gitignore`, but I think we can just check this file in for simplicity. This also simplifies the development process a little bit since if the `pyrefly_wasm_bg.wasm` already exists locally, then I don't need to rebuild the wasm during development.

Opening this up for RFC in case I missed a downside of doing this

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D72580838

fbshipit-source-id: adfb611b29a1ef8a16f232ce0afcbfb54ebb4b63