Prepare `Server.send` to be `Send + Sync + 'static`
Summary:
When we implement parallel recheck and IDE services, the recheck will happen in another thread. At the end of the recheck thread, we need to send diagnostics. The send function must satisfy `Send + Sync + 'static` so that we can pass it onto the recheck thread that can out live the stack.

This diff does the prep work to make the `send` function satisfies the constraint, mostly through wrapping the connection in an Arc.

Reviewed By: ndmitchell

Differential Revision: D72675658

fbshipit-source-id: 08a663263eeaad51cacb390a3792cdb5d568cc2c