Flatten `narrow` and `narrow_type`
Summary:
In order to begin working on the internals of solver-side narrowing,
I had to introduce a wrapper around the main narrow function because
we were recursively calling into it from lower-level places, which made
it impossible to make the outer function operate at the `TypeInfo` level
cleanly.

We've eliminated those recursive calls (and along the way simplified
the `NarrowOps` data types and fixed a bug on isinstance-union), which
means we can finally make the `narrow` function actually operate at
the `TypeInfo` level now.

The only immediate behavior change is that an atomic narrow operation
would preserve existing attribute narrows if there were any; the change
isn't visible yet since we still have no way to actually create attribute
narrows but the logic should be correct relative to upcoming functionality.

Reviewed By: ndmitchell

Differential Revision: D72718431

fbshipit-source-id: b280085ea11936cee0be30a51869c40a8ad2cf76