Implement the most basic attribute narrowing
Summary:
We're finally ready to actually wire attribute narrowing up to the solver.

To get this off the ground, I put in the most basic possible implementation, where
we've ignored completely the question of what type might be there before the narrowing:
we just assume `object` which is mostly sound and ignoring special cases like descriptor
access for the moment.

All three tests pass - we can't do much with the narrows yet, but narrowing seems to be
correctly filling out the attribute tree in `TypeInfo`.

Reviewed By: ndmitchell

Differential Revision: D72718427

fbshipit-source-id: 310b729b3120dc90bf90213622236e61d6b8d567