Patch up the type on `add_narrow`
Summary:
I'm finally ready to use this, and I figured out that passing a vec of
`&Name` is definitely a bad idea - the names are actually going to be
coming from a datum embedded inside a `NarrowOps` that lives inside a
`Binding`, which means they are going to be owned (as a Vec1) when they
get to us, which in turn means we really just want to pass along that Vec1
rather than making a new vec of references.

What we need to do is pass a reference to the whole thing, not a vec of references.
Unfortunately this does seem to prevent me using the safe `split_off_first` API,
which requires owning the Vec1, but `split_first` works fine (we just have to
manually mark the failure branch as unreachable).

Easy fix; to make the test code non-brittle (it really gets annoying to have
to call `clone()` on all but the last occurence of a local var) I just made the
helpers in tests into little closures that produce owned data on every
invokation.

While I'm at it, make the mutable api private; I realized we're going to need
a more functional `with_narrow` at the callsite, and the tests are local, so
there's no reason to make this `pub`.

Reviewed By: ndmitchell

Differential Revision: D72718429

fbshipit-source-id: 50238530413501c6895078df8e32fe4a359b43cf