Center align header on pyrefly.org and have it fill the entire screen
Summary: On the old landing page, the header is currently not vertically centre aligned and clips to the top. As a result, there's white background shown at the bottom of the page if you were to zoom out. This diff fixes this by making sure that the header takes up the entire space of the layout.

Reviewed By: migeed-z

Differential Revision: D72723948

fbshipit-source-id: 3a84a865f5834a04d4d22a51d2dbcf9d7a16cecc