Try to narrow on `self.assertTrue()` and `self.assertFalse()` in tests
Summary:
This diff aims at supporting refinements when `unittest.TestCase.assertXXX` methods. The typical code pattern we see is as follows:
```
# some test file
import unittest
def foo() -> MyClass | None: ...
class MyTest(unittest.TestCase):
  def test_derp(self) -> None:
    x = foo()
    self.assertIsNotNone(x)  # Or some other forms of assertions
    # use `x` as if it has type `MyClass`
```
The basic approach here is to treat the `self.assertTrue(XXX)` method invocation as if we are seeing an actually assert statement `assert XXX`, where a narrowing op gets inserted into the binding table.

This is mostly straightforward, except for one issue: how do we know if `self.assertTrue()` is really invoking `unittest.TestCase.assertTrue()`? In principle, the correct way to do it would be to grab the type of `self.assertTrue` during solving and asks where this type comes from. But given that we are already not doing the principled thing for `unittest.TestCase` in D70978112, I figured I should at least keep things consistent with how that's done :P

Reviewed By: ndmitchell

Differential Revision: D72662392

fbshipit-source-id: 14d0699286f1b1399e5155e2094590d07f13ff89