Create ModuleGlob
Summary:
While working with Teddy and Zeina, we determined it would be better if values in `replace_imports_with_any` could be regex values. This matches how mypy works, and some of the projects we're working on expect those values to be regex instead of prefix-matched.

We ran into a regex that looks like this, which motivates us to make this change `path.to.some.*.fb`.

This diff creates a `ModuleGlob` type with basic functionality that we'll be able to deserialize values directly into when parsing a config file. In the next diff we add the functionality that converts it from a module glob to a regex.

Reviewed By: ndmitchell

Differential Revision: D72432740

fbshipit-source-id: 8f8ac911bf8426d70a94185966dd1fc1c4a4a332