fail github action if version can't be found
Summary:
we were green on all our github when the version couldn't be found in deploy_extension
{F1976941572}

this has caused our vscode extension to release using the default version (`1.0`) which means any newer versions (`0.10.0`) will be "older" and not used. these actions should be red if we can't find the version.

Reviewed By: ndmitchell

Differential Revision: D72737401

fbshipit-source-id: 809cb370fdea4dcf324180b113b51fff550b0f4e