set version to 0.0.1
Summary:
if anyone accidentally runs `npx vsce publish` without `--version <version from version.bzl>`, it will use the version in package.json. it is especially annoying if this published version is higher than the deployed version because users will get the highest version in the store, not the newest version.

this version should be set to 0.0.1 so a pushed version is unlikely to be higher than the newest version.

see [this post](https://fb.workplace.com/groups/657557118131740/permalink/1712594672627974/) for more details

Reviewed By: ndmitchell

Differential Revision: D72738812

fbshipit-source-id: 5cff19e6f7f433a55244ef77b6cdb340b32ce1b6