fix security vulnerabilities
Summary:
allow-large-files

Solumin fixed a vulnerability task as the oncall last week: T219960014. That got me curious and I wanted to see how many other vulnerabilities. Turns out there was already 10 in the Pyrefly repo. This diff cuts it down to 1.

I had devmate generate a script (`audit_and_fix_vulnerabilities.sh`) to address the vulnerabilities, which runs `yarn upgrade` and then checks how many scripts are remaining. I then had to manually upgrade one of the packages to address a severe vulnerability. There's one more moderate vulnerability which doesn't seem to be addressable since we are already on the latest version of `docusaurus-plugin-internaldocs-fb`. I would expect that when a new version is released, then this would be fixed

Reviewed By: Solumin

Differential Revision: D72672898

fbshipit-source-id: 1443ac989e0c035ffd9adb3707aa48d6e58903b2