exclude pyrefly_wasm implementation to unblock docusaurus build
Summary: Maggie mentioned offline that her development using on demands broke because of D72580838. I wasn't even aware that people are using on demands for development. Reverting this change to unblock on demands. If there are help from other people that will help us with the website, it's possible that they will be using on demands as well

Reviewed By: MaggieMoss

Differential Revision: D72748516

fbshipit-source-id: c3d4c5e20c1502775b9740e3bfdb888d48f62e4b