Add new helper `at_name` to `TypeInfo`
Summary:
This simple helper treates the narrows as an overlay against a fallback
lookup, which is only called if we need to compute a type for the root.

It preserves any subtree of narrowed attributes, whether or not the
root at this name had a narrow.

It is not yet used, but will be used to power `attr_infer`.

Reviewed By: ndmitchell

Differential Revision: D72793257

fbshipit-source-id: d4db74f70004892948fdf299fd5f7ef343cde8cb