Stop assuming __eq__ always returns a bool
Summary:
I noticed that some of the errors we produce in pandas are due to us assuming a `==` comparison returns bool.

I also had to fix up rich comparison method generation for dataclasses - we were generating the wrong signatures for `__eq__` and `__ne__`, which was hidden by the special-casing of `__eq__` in operators.rs.

Reviewed By: ndmitchell

Differential Revision: D72767027

fbshipit-source-id: d9efaa5b82a9e4ba5e941b0ee9dea12106c692fa