Track a bug in comprehension control flow
Summary:
As I was working on attribute narrowing, I hit a test case (the
previous version of `test_listcomp_if_error`) that broke.

I realized by experimenting (leading to `test_listcomp_if_control_flow`)
that the issue was not attribute narrowing per se, but rather a pre-existing
bug in narrowing control flow: in comprehensions, we appear to be
applying the narrow from the `if` *before* binding names in the expression
for type check.

This leads to missing errors if the narrow masks a type error that isn't
actually guarded; for attribute narrows the issue is especially salient
because we can narrow on a nonexistant attribute and we'll never see the
error about no such attribute existing.

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D72800927

fbshipit-source-id: 255fc2ded0eb8ff80e9408c688683eeed7fad974