Preserve function metadata on callback protocols
Summary:
When a decorator turns a function into a callback protocol, we need to convert it back to a function so we can attach metadata to it (like whether it's a property).

This would regress a conformance test if we didn't preserve information about what other attributes are available on the protocol, so I also had to tweak attribute lookup.

Fixes https://github.com/facebook/pyrefly/issues/85.

Reviewed By: ndmitchell

Differential Revision: D72844954

fbshipit-source-id: 8d59c2a3d3d9db502106b671a194f90e23deb047