Wire up attribute reads in `attr_infer`
Summary:
Now narrowed reads work end-to-end within an expression.

I extended the test a bit to cover a few straight-line cases; it's a
little limited how much we can do in practice because at this stage
we still don't propagate narrowing information through assignments.

Note that there's a more realistic test whose output changed but is
still buggy - this is because at the moment the attribute narrowing
itself and the lookups both work, but we haven't yet wired up the
attribute reads into narrowing so we always start at `object`.

That means negative narrowings like `is not None` narrowing never actually
work, we just get `object` back out.

Now that we support reads, I'll work on using read results in narrowing
next; without this the narrowing won't benefit us much in real code, since
`is not None` is by far the most common actual narrow.

Reviewed By: ndmitchell

Differential Revision: D72793140

fbshipit-source-id: b70dec89f56076ab1ebc8e539cf42151c3933dfd