Implement basic attribute-lookup-for-narrow
Summary:
This diff introduces the logic to look up an attribute's prior type in
narrowing and use that type in the narrowing type algebra, which is needed
to make negative cases work (in particular the all-important `is not None`
case).

Previous test cases pass, and I added a new test verifying the behavior
when the attribute read wasn't valid to start with - in that case we
error at narrow time but still apply the narrow (IMO this is the only
acceptable option for gradually typed code - the narrow might be acting
as a sort of cast and we should respect it downstream).

-----------------

The API is intended to allow us to track when the user has narrowed a
descriptor read, but I'm putting that off for now - it is not actually
possible to do well today because the text ranges for narrows overlap,
so we would get many duplicate errors. We probably want to actually store
information in `TypeInfo` and warn downstream anyway because certain
unit test use cases (basically any assert on a property read!) would become
frustrating in strict mode if we are too eagerly complaining.

I'll likely get to that in a couple weeks at soonest, it will be a chunk of
work but is not fundamentally hard and is unlikely to materially impact the
architecture of Pyrefly. It is not needed for parity with Pyright or Mypy
because both of them unsoundly narrow:
 - [pyright](https://pyright-play.net/?code=JYWwDg9gTgLgBFAhgOwCYRAKAGZQ3GATzGGQHM5RJYEBTAN1sQBsB9IsWzTAY2cQDOAuADEIEAFyY4MuAAEweTrELTZqWtjgAPABQDazbAEo4AWgB8cATChwAPnAByEZLSmzPdGAFcoyOAAibUDKLV0kNAwAOkj0EF1TAB44AAZogFZTQwNnVy5MDS1scV0SyVFxYw9ZYGLxaO1KYWQIeBc3Gq86RhZ2Yloyhu1jTCA)
 - [mypy](https://mypy-play.net/?mypy=latest&python=3.12&gist=d86ddf1022bf24397615b97cebcf6e74)

But it might be a pain point for the most strictness-loving Pyre users;
[Pyre refuses to narrow](https://fburl.com/j07bvg5e).

Reviewed By: ndmitchell

Differential Revision: D72817298

fbshipit-source-id: 96fee25c9fd17162dfb458d7288b51e1d13b7504