Add test cases for top level type ignore
Summary:
I think there's a bug in the existing implementation of #pyrefly: ignore

It thinks the new flag I added, pyrefly: ignore-all-errors is a pyrefly: ignore, since it's doing prefix matching.

It seems we should fix that bug first before implementing the flag (unless we name the top level flag #pyrefly-ignore-all or something.

Not sure how easy the fix is, but I can look into it next

Reviewed By: ndmitchell

Differential Revision: D72881340

fbshipit-source-id: 21047da36dde5e8e1334861f27ab27b8ac0f8d6d