Begin implementing top level type ignore
Summary:
I think we should still not consider any comment that starts with pyrefly: ignore-all as a line level flag. But in this diff, we begin implementing a top level flag to ignore all errors in a file.

In this diff, we add a new flag that determines if a given comment is top level.

If we discover a top level comment that ignores all errors, then we record that and ignore the whole file.

RFC: should we add any suppressions at all when seeing a top level comment? My guess is no. So I wonder if we should just stop the loop that collects suppressions when we encounter a top level ignore.

Reviewed By: ndmitchell

Differential Revision: D72883343

fbshipit-source-id: 928decf48c36e875a5844107e55305450f84ac4f