Pass transaction to each IDE command
Summary:
Later in the stack when we implement parallelize all IDE operations and recheck, it's no longer the case that the transaction we use for these IDE operations are fresh from the main `State`, since during recheck, we will temporarily hold the latest in-memory content in a non-committable transaction.

This diff prepares for that by passing a transaction to these IDE operation functions.

Reviewed By: ndmitchell

Differential Revision: D72872212

fbshipit-source-id: 0b66042d3721f327e96b1a4ccee7ac0ae918813c