Add `State.try_new_committable_transaction` API
Summary: This diff adds an API to `State` where we can optimistically request a committable transaction. This is helpful in the IDE use case, where we will try to use committable transaction for in-memory file changes, but if that's not possible during an ongoing recheck, it's better to not block and temporarily switch to use a non-committable transaction to hold the latest content.

Reviewed By: ndmitchell

Differential Revision: D72872209

fbshipit-source-id: e818b008fb2c1a15149cf5646a1c7814e85801b5