Force variables in union
Summary: We often do map over union, then assume the inside is not a union. That fails if the Var points at a union, so we instead try and expand the union first. Fixes a few bugs, but defaults them to Any rather than int, so will have to add test cases that show things are still broken next.

Reviewed By: rchen152

Differential Revision: D72921282

fbshipit-source-id: 64ddff741529650adb2211acefaec065ee99f83c