Shift some read operations on `ReadableState` into `Transaction`
Summary:
Later in the stack when we implement parallelize all IDE operations and recheck, it's no longer the case that the `State` will always contain the latest data for IDE operations, since during recheck, we will temporarily hold the latest in-memory content in a non-committable transaction.

This diff prepares for that by shifting some read operations `ReadableState` into `Transaction`, so that we can always read the latest data for IDE operations, even during recheck.

Reviewed By: ndmitchell

Differential Revision: D72877478

fbshipit-source-id: e7c5704635aa09690ae3189b8d94bab7aa550092