Split `Transaction` into `TransactionSavedState` and `Transaction`
Summary:
Later in the stack when we implement parallelize all IDE operations and recheck, during recheck, we will temporarily hold the latest in-memory content in a non-committable transaction. However, doing so will keep the hold of the read lock of `ReadableState` (even when we are not actively doing IDE operations), which will block the commit of recheck.

We can resolve the problem by splitting `Transaction` into `TransactionSavedState` and `Transaction`, which can convert between each other. When we are not doing reads or checking, but still want to hold some uncommitted changes (e.g. during recheck but we still want to validate in-memory content), we can switch to `TransactionSavedState`, and when we want to check or read, we can switch to `Transaction`.

Reviewed By: ndmitchell

Differential Revision: D72872210

fbshipit-source-id: 60e88637b0e3eb4361cb03f73c8245415b46913a