Add a default to every variable in a loop
Summary:
Given:

```
n = 0
while cond:
    n += 1
```

You end up with the bindings:

```
n0 = 0
n1 = phi(n0, n2)
n2 = [[ n1 += 1 ]]
```

We used to attach a default to the `n1` node. Which was non-deterministic, because if you arrived at `n2` first then that would cause the recursive loop, and it lacked a default. Solve this by adding a default to every assignment inside the loop, which hits both `n1` and `n2` and makes the result stable. Before it varied based on whether you did an `assert_type` inside the loop or not.

Reviewed By: rchen152

Differential Revision: D72933865

fbshipit-source-id: 22e32b27b8c4b138662503df5a8982e55c0633fa