Parallelize all IDE operations and recheck
Summary:
With all the preps before, this is the diff where we parallelize all IDE operations and recheck.

First, we will keep the happy path the same: when there is no ongoing rechecks, we will continue to use committable transaction for in-memory changes. This is enabled by the API I added in D72872209.

Now for the complicated case: what should we do when we failed to get a committable transaction (i.e. there is an ongoing recheck). In that case, we will use an uncommitable transaction to hold the latest in-memory changes. Naively doing so can cause dead lock, so in D72872210 I split the transaction into `Transaction` and `TransactionSavedState`, so that we don't have to hold the read lock on `ReadableState` when we are not doing anything.

Of course, we want to eventually get off the temporary uncommitable transaction once recheck finishes. This is done at the first `validate_in_memory` call when we discovered that there is no ongoing recheck.

With these changes, all IDE operations are fully parallel with rechecks (except for the committing stage that's quick).

Reviewed By: ndmitchell

Differential Revision: D72872213

fbshipit-source-id: 649e27783ff6a1448b5290e12e85ffc40d6475be