Emit an error and mark a class as inheriting from Any when we see an invalid base class
Summary:
This fixes a few attribute errors in pybind11. It also adds a couple of invalid-inheritance errors, but those seem more legitimate =P

It's a bit debatable whether the test I added contains any error at all, but (1) pyright and mypy both report an error on the class definition, and (2) we can't handle a union base class, so better to warn the user than silently convert it to Any.

Reviewed By: ndmitchell

Differential Revision: D73019521

fbshipit-source-id: 0f2a0c19cf7ad9136d8e2faf8eb7f6bc4395b143