Fix up loop defaulting
Summary: Previously when ending a flow (if or loop) we moved the default from who knows where based on a completely broken understanding of what we were looping over (see previous diff...). We should instead use the existing default, unless we are not in a loop, in which case we should reset it.

Reviewed By: rchen152

Differential Revision: D73062945

fbshipit-source-id: 67875ca5e0e6115e7f987f4ef006a142eab7d172