Add script to build wasm for testing
Summary:
I want to add a CI test for wasm, to be able to ensure that the basic IDE functions such as getting errors, go-to-def, autocomplete, etc all works as expected. I tried using the existing wasm file, but I ran into issues when integrating with jest since jest doesn't play well with [es6](https://www.w3schools.com/js/js_es6.asp). I tried a couple different thing, but none seem to work. The only working example seem to be doing a separate wasm build with a node target (based on the post [here](https://stackoverflow.com/questions/65710914/is-it-possible-to-test-wasm-pack-generated-webassembly-from-a-javascript-context)).

This diff sets up a separate build specifically for tests. See following diffs for how this will be used.

Reviewed By: SamChou19815

Differential Revision: D72980529

fbshipit-source-id: a8a45b1b0bebcc4b5ee531424bd62b274aa3eecc