Add basic wasm test
Summary: Adds a basic wasm test by verifying that we are able to directly call into wasm and see that it returns the errors that we expect. This diff also sets up helpers that's necessary to fetch the correct wasm file we build using `build-wasm-for-test`. We'll add more complex tests in the following diff (D72980530).

Reviewed By: stroxler

Differential Revision: D72474171

fbshipit-source-id: ed5731ede19ceba4ce372a35fcffeceebf384838