Fix stale error issue happening during current recheck and unsaved edits
Summary:
Currently, there is a race condition that can cause stale errors when we have current recheck and unsaved edits. It has two causes:

1. When the recheck finishes, it will publish new diagnostics. However, the errors are coming from potentially stale in-memory content.
2. We don't send new diagnostics when recheck is ongoing.

To address the issue, we slightly revamp when and where we send diagnostics. We will only send diagnostics on the main thread based on the fresh in-memory content, and potentially stale main `State`. This can happen regardless of whether we can commit the transaction when validating in memory content. When recheck finishes, instead of publishing the errors on its own, it will send a `RecheckFinished` event to the server main event loop, so that the server on the main thread can immediately revalidate in-memory content with the fresh main `State`.

Reviewed By: kinto0

Differential Revision: D73041758

fbshipit-source-id: df43d3b2a3e0e7a8e9ad551f3bbf1624436f400e