Make all Config options option types
Summary:
# This Feature...
Over the next few diffs, we'll be implementing the functionality for sub-configs. This involves selecting configuration values from a sub-config for a matched file path, overriding the value used in the root config (if present). It will allow users with multiple projects in the same repository that should use different settings to work under the same config, type checking everything at once, and could make LSP configuration easier if we choose to emulate Pyright's approach.

## High Level Approach

1. Make all config values `Option` types. This will let us know that we need to fall back to the next sub-config or the root config, instead of taking whatever default value we find.
2. Make a `BaseConfig` type, which will be flattened into the main `ConfigFile` type.
3. Make a `Glob` type (pull some of the logic from `Globs`).
4. Create `SubConfigs` and add them to `ConfigFile`, which are `BaseConfigs` paired with a `Glob`.
5. Implement configuration resolution, which will figure out which `SubConfig` to use (or the default), and grab the given config value from it.

# This Diff...

Makes the rest of the top-level config items in a `ConfigFile` `Option` types, and adds getters for them.

As with the previous diff, the getters here should do their best to return references, and can safely use `unwrap()` for config values on the root config (we dynamically check for all option types to not be `None` on the root config`.

Reviewed By: ndmitchell

Differential Revision: D72982649

fbshipit-source-id: b7dcb421805cb36e7bb00bb5f2e41be1b38898c7