Skip warning on nonexistent site package path if provided by interpreter query
Summary:
We should avoid warning on a nonexistent site package path if it was one provided by an interpreter, since sometimes they might not exist. This diff adds a new hidden field to `PythonEnvironment`, which we set to true when getting one from an interpreter query.

Addresses https://github.com/facebook/pyrefly/issues/78

Reviewed By: ndmitchell

Differential Revision: D73083558

fbshipit-source-id: b70051a8f8db334b46c062ff2dae97c828b34539