Create SubConfig (#91)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/91

# This Feature...
Over the next few diffs, we'll be implementing the functionality for sub-configs. This involves selecting configuration values from a sub-config for a matched file path, overriding the value used in the root config (if present). It will allow users with multiple projects in the same repository that should use different settings to work under the same config, type checking everything at once, and could make LSP configuration easier if we choose to emulate Pyright's approach.

## High Level Approach

1. Make all config values `Option` types. This will let us know that we need to fall back to the next sub-config or the root config, instead of taking whatever default value we find.
2. Make a `BaseConfig` type, which will be flattened into the main `ConfigFile` type.
3. Make a `Glob` type (pull some of the logic from `Globs`).
4. Create `SubConfigs` and add them to `ConfigFile`, which are `BaseConfigs` paired with a `Glob`.
5. Implement configuration resolution, which will figure out which `SubConfig` to use (or the default), and grab the given config value from it.

# This Diff...

This diff creates the `SubConfig` struct, which allows us to store `SubConfig`s in a `ConfigFile`. The main changes involve adding the field to the `ConfigFile` struct, updating `from_file()` and `validate()` to operate on sub configs in addition to the root, and adding the field to struct construction when required.

The rest of the changes add tests to validate the behavior works as expected.

Reviewed By: ndmitchell

Differential Revision: D73000284

fbshipit-source-id: a45803a0b4fc9da74f0e473d467ca93495f31159