Handle empty join in TypeInfo
Summary:
It is not possible to get an empty join from `or` handling, so I didn't
catch this in my initial implementation even with an attempt to stress
test all cases.

But it *is* possible to get empty joins from `Phi` nodes (due to overload
filtering - otherwise it would not be).

I thought split_off would just produce two empty vecs, but it actually panics
so we need to guard against that explicitly in `NarrowedAttrs` join; adding
a unit test for this since it's pretty easy to construct the data manually
and it's not very easy to construct an actual Python example that triggers
this case.

Reviewed By: grievejia

Differential Revision: D73048080

fbshipit-source-id: c457b827fd1a1fa13165df983873900d7e2849e5