Implement TypeInfo::join, support `or` narrowing
Summary:
This diff implements `TypeInfo::join`, which is a textbook map join operation.
We already wired it up to the narrowing code, so now `or` narrowing works.

I've only got a single test in place ahead of time but the next diff tests all the
edge cases I could think of.

--------------------

The implementation here points to some potential simplifications of `TypeInfo`
data ... I think the `NarrowedAttr` enum (which tries to express and invariant
that any node should have at least one oof a type or subtree) isn't really
benefitting us much, and the option should probably be moved outside of
`NarrowedAttrs` instead of living inside.

I'm going to leave the data structure alone for now - it seems better
to keep working on functionality first, especially since the more functionality
and tests we have the more confident we can be in a refactor - but there are a
few places where we could clean the join code up if we tweak the datatypes.

Reviewed By: ndmitchell

Differential Revision: D73048081

fbshipit-source-id: 0008e891a985180638d29f805d64053e98df268e