Demonstrate that control flow does not break narrows
Summary:
This diff adds a test showing that control flow structures do not interrupt attribute
narrows. This is nontrivial; I actually was expecting that the copious Phi nodes we
create might lead to bugs because we don't yet handle narrows on Phi; I think this
works because we optimize the no-bind case to a Forward and that already propagates.

Reviewed By: ndmitchell, grievejia

Differential Revision: D73048077

fbshipit-source-id: 91e315c35d96f5804b2013f60cb3412e5f0b6997