Use `TypeInfo::join` for `Phi` nodes
Summary:
Needed for attribute narrowing to correctly work with control flow.

-----------

The tests all behave as expected in that we join attribute narrows.

But it also exposed that `self.unions` isn't currently simplifying
linearly orderable types, which feels like a gap (and could certainly
make our hover / `reveal_type` results a bit confusing in the presence
of narrows).

That's not actually an attribute-narrowing-specific issue, so I'm not
tackling it now but I did mark those spots with a `TODO` so we remember
to come back to this at some point.

Reviewed By: ndmitchell, grievejia

Differential Revision: D73048076

fbshipit-source-id: 979f4c7960d03bf7e640ba7dbe3043286c3fd3e1