Expand `or` tests to cover most join logic
Summary:
I initially put a very simple `or` test in place and let the types guide
my full implementation of `join`.

But we ought to test most of the known edge cases (folding subtrees,
dropping non-overlapping narrows, etc) in unit tests; this diff expands
the `or` tests considerably to cover the cases I could think of. No bugs
found so far, luckily the types are a pretty good guardrail :)

Reviewed By: grievejia

Differential Revision: D73048082

fbshipit-source-id: eee01aa206d2be14da360b468afb7e000c43c11a