use disableLanguageServices flag
Summary:
see previous diff for context. this diff conditionally disables language services (go-to def, hover, completions, inlay hints) based on the config vscode sends us.

because it's in the vscode config, it will be re-queried on didChangeConfiguration. didChangeConfiguration is sent again on changes to this value so no window reload is necessary for these changes to take effect.

note: when disabled, we fast-return `[]` for go-to definition instead of `None` like [pyright does](https://fburl.com/code/y133lty4). the difference between `[]` and `None` has been debated (partially relevant post [here](https://fb.workplace.com/groups/159208099316174/permalink/464360585467589/)) and there's no clear reason to use one or the other so i'm leaving it as is for now. if people feel strongly (I don't) or we think the difference is necessary for metrics we can change it at some point.

Reviewed By: SamChou19815

Differential Revision: D73063436

fbshipit-source-id: 42602386445ffff9504d376b3b9144f13ac0237d