Accept more valid arguments in check_type_is_class_object
Summary: This brings our conformance results almost back to where they were before D73070094 (456 failures instead of 454). The two extra failures are due to a check we don't implement yet for data protocols in issubclass calls; we previously accidentally "passed" these conformance tests because we errored on the issubclass calls for other reasons.

Reviewed By: grievejia

Differential Revision: D73075386

fbshipit-source-id: afdf8687c7f6ef3de242048aefe0cf43c9523d63