Don't lose Type::Type wrapper when decomposing unions for narrowing
Summary: I noticed while reviewing D73064910 that we have a surprising Type::ClassType(...) branch in unwrap_class_object_silently. (Type::ClassType(...) is an *instance* of a class, which shouldn't be a class object.) Turns out we needed it because we were dropping Type::Type along the way while dealing with unions. Add the lost Type::Type back and delete the surprising branch.

Reviewed By: grievejia

Differential Revision: D73067830

fbshipit-source-id: 719fa45292277202afa0010bc18fb7a3e2c2d93a