improve NamedTuple field checks
Summary:
Based on some testing in the CLI
- methods are never counted as fields (but lambdas work, unlike enums)
- it's a runtime error if named tuple fields start with an underscore, with some caveats
  - if it's a method, then it's not a field so there's no runtime error if it starts with underscore.
  - if a field is renamed to _0, _1 using the functional syntax rename=True param, then there's no error if it starts with underscore
  - if a user-written field starts with an underscore and rename=True is set, it will be renamed to _0, _1, etc.

This diff implements all of the above functionality, EXCEPT disallowing fields starting with underscore for the class syntax. This is because we don't have a way to differentiate a renamed field from a field originally written by person.

I think we could probably do something about this by moving every field on a synthesized class into the same structure that we're using for synthesized methods, but that's a bigger change for a separate diff. For now, I left a testcase with some comments.

fixes https://github.com/facebook/pyrefly/issues/101

Reviewed By: stroxler

Differential Revision: D73113784

fbshipit-source-id: 2f20e6b949e2e17dbfedf36f153d333cae98a786